copy_svg_langs_per_file > copy_svg_langs
