+++ src/main_app/jobs_workers/public_jobs_workers/copy_svg_langs/steps/__init__.py
@@ -1,17 +1,12 @@
-from .download_step import download_step
 from .extract_text import extract_text_step
 from .extract_titles import extract_titles_step
 from .extract_translations import extract_translations_step
-from .fix_nested import fix_nested_step
-from .inject import inject_step
-from .upload import upload_step
+from .inject_one_file import InjectResult, inject_step_one_file
 
 __all__ = [
-    "download_step",
     "extract_translations_step",
-    "fix_nested_step",
-    "inject_step",
+    "inject_step_one_file",
+    "InjectResult",
     "extract_text_step",
     "extract_titles_step",
-    "upload_step",
 ]
