+++ src/main_app/jobs_workers/public_jobs_workers/copy_svg_langs/worker.py
@@ -9,23 +9,29 @@ import logging
 import re
 import threading
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any, Tuple
 
 import requests
 from mwclient.client import Site
 
 from ....api_services import create_commons_session, get_user_site
+from ....api_services.files_service import download_svg_file, upload_fixed_svg
 from ....config import settings
+from ....shared.fix_nested.worker import (
+    DetectionResult,
+    VerificationResult,
+    detect_nested_tags,
+    fix_nested_tags,
+    verify_fix,
+)
 from ...base_worker_object import BaseObjectsJobWorker
-from .objects import CopySvgLangsWorkerObject, FilesProcessedItem, StageDetail, StepResult
+from .objects import CopySvgLangsWorkerObject, FilesProcessedItem, FileSteps, StepResult
 from .steps import (
-    download_step,
+    InjectResult,
     extract_text_step,
     extract_titles_step,
     extract_translations_step,
-    fix_nested_step,
-    inject_step,
-    upload_step,
+    inject_step_one_file,
 )
 
 logger = logging.getLogger(__name__)
@@ -69,10 +75,6 @@ class CopySvgLangsWorker(BaseObjectsJobWorker):
         self.main_title: str = ""
         self.titles: list[str] = []
         self.translations: dict[str, str] = {}
-        self.nested_data: dict[str, str] = {}
-        self.nested_results: dict[str, str] = {}
-        self.inject_data: dict[str, Any] = {}
-        self.files_to_upload: dict[str, str] = {}
 
     def get_job_type(self) -> str:
         """Return the job type identifier."""
@@ -103,26 +105,6 @@ class CopySvgLangsWorker(BaseObjectsJobWorker):
         out_dir_main.mkdir(parents=True, exist_ok=True)
 
         return out
-
-    def _is_cancelled(self, stage: StageDetail) -> bool:
-        if self.is_cancelled():
-            stage.status = "cancelled"
-            return True
-
-        return False
-
-    def log_upload_error(self, _msg, result, status) -> None:
-        self.result.stages.upload.status = status
-        self.result.stages.upload.message = _msg
-
-        for _, item in self.result.files_processed.items():
-            if item.status != "failed":
-                item.steps.upload = StepResult(
-                    result=result,
-                    msg=_msg,
-                )
-                item.status = status
-                item.error = _msg
 
     def _save_files_stats(self, stats_data) -> None:
         files_stats_path = self.output_dir / "files_stats.json"
@@ -169,54 +151,6 @@ class CopySvgLangsWorker(BaseObjectsJobWorker):
             self.titles = list(step_result["titles"])
             self.titles.sort()
 
-            return True
-
-        stage.status = "failed"
-        stage.message = step_result.get("error", "Unknown error")
-        self.result.status = "failed"
-        return False
-
-    def _run_stage(
-        self,
-        stage: StageDetail,
-        step_func: Callable[[], dict[str, Any]],
-        run_after_func: Callable[[], None] | None = None,
-    ) -> bool:
-        """Run a single stage and update result."""
-        stage_name = stage.name
-
-        if self.is_cancelled():
-            stage.status = "cancelled"
-            return False
-
-        stage.status = "running"
-        self._save_progress()
-
-        try:
-            # Call the lambda / callback directly without passing args
-            step_result = step_func()
-
-        except Exception as e:
-            logger.exception(f"Error in stage {stage_name}")
-            stage.status = "failed"
-            stage.message = str(e)
-            self.result.status = "failed"
-
-            return False
-        stage.data = step_result
-
-        if step_result.get("message"):
-            stage.message = step_result["message"]
-
-        if step_result.get("success"):
-            stage.status = "completed"
-            if "summary" in step_result and not stage.message:
-                summary = step_result["summary"]
-                if isinstance(summary, dict):
-                    stage.message = ", ".join(f"{k}: {v}" for k, v in summary.items())
-
-            if run_after_func:
-                run_after_func()
             return True
 
         stage.status = "failed"
@@ -286,35 +220,6 @@ class CopySvgLangsWorker(BaseObjectsJobWorker):
 
         return False
 
-    def save_stats_summary(self) -> None:
-        stats_data = {
-            "main_title": self.main_title,
-            "translations": self.translations,
-            "titles": self.titles,
-            "files": self.files_dict,
-            "files_dict": self.files_dict,
-            "nested_task_result": self.nested_data,
-            "injects_result": self.inject_data,
-        }
-
-        self._save_files_stats(stats_data)
-
-        # Compile final results for database
-        self.result.results_summary.update(
-            {
-                "total_files": len(self.files_dict),
-                "files_to_upload_count": len(self.files_to_upload),
-                "no_file_path": len(self.files_dict) - len(self.files_to_upload),
-                "injects_result": {
-                    "nested_files": self.inject_data.get("nested_files", 0),
-                    "success": self.inject_data.get("success", 0),
-                    "failed": self.inject_data.get("failed", 0),
-                },
-                "new_translations_count": len(self.translations.get("new", {})),
-                "main_title": self.main_title,
-            }
-        )
-
     def process(self) -> CopySvgLangsWorkerObject:
         """Execute the full pipeline."""
         if not self.title:
@@ -324,14 +229,8 @@ class CopySvgLangsWorker(BaseObjectsJobWorker):
 
         self.site = get_user_site(self.user)
         if not self.site:
-            upload_result = {"done": 0, "not_done": len(self.files_to_upload), "failed": True}
-            self.result.results_summary["upload_result"] = upload_result
-            self.log_upload_error("Authentication failed", False, "failed")
+            logger.warning("Job %s: No site authentication available", self.job_id)
             self.log_no_site_error()
-            # ----------------------------------------------
-            # Stage 8: save stats and mark done
-            self.save_stats_summary()
-
             return self.result
 
         self.session = create_commons_session(settings.other.user_agent)
@@ -358,213 +257,263 @@ class CopySvgLangsWorker(BaseObjectsJobWorker):
         if not self._extract_translations_step():
             return self.result
 
-        # ----------------------------------------------
-        # Stage 4: download SVG files
-
-        def download_progress(index: int, total: int, msg: str, results: dict[str, Any]) -> None:
-            self.result.stages.download.message = f"Downloading {index}/{total}: {msg}"
-            # if index % 10 == 0:
-            self._save_progress()
-
-            for title, _result in results.items():
-                if title in self.result.files_processed:
-                    item = self.result.files_processed[title]
-                    if not item.steps.download.msg:  # to avoid overwriting previous messages
-                        item.steps.download = StepResult(
-                            result=_result.get("result", ""),
-                            msg=_result.get("msg", ""),
-                        )
-                        if _result.get("result", "") is False:
-                            item.status = "failed"
-                            item.error = _result.get("msg", "")
-
-        def download_run_after() -> None:
-            download_data = self.result.stages.download.data
-            self.files_dict = download_data["files_dict"]
-
-            # clean up
-            self.result.stages.download.data["results"] = {}
-            self.result.stages.download.data["files"] = []
-
-        # Initialize files_processed
+        processfiles_stage = self.result.stages.processfiles
 
         title_to_work = self._apply_limits(self.titles)
 
-        for title in title_to_work:
-            self.result.files_processed[title] = FilesProcessedItem(
+        per_item = self.get_priority(len(title_to_work))
+        processfiles_stage.status = "running"
+
+        for n, title in enumerate(title_to_work, start=1):
+            processfiles_stage.message = f"Processing files {n}/{len(title_to_work)}"
+            logger.info("Job %s: Processing title %d/%d: %s", self.job_id, n, len(title_to_work), title)
+
+            if self.is_cancelled():
+                logger.info("Job %s: Cancellation detected, stopping.", self.job_id)
+                processfiles_stage.status = "cancelled"
+                break
+
+            title_info = FilesProcessedItem(
                 title=title,
+                file_path=None,
                 status="pending",
                 error=None,
+                steps=FileSteps(
+                    download=StepResult(result=None, msg=""),
+                    nested=StepResult(result=None, msg=""),
+                    inject=StepResult(result=None, msg=""),
+                    upload=StepResult(result=None, msg=""),
+                ),
             )
+            ok = self._process_one(title, title_info)
 
-        if not self._run_stage(
-            self.result.stages.download,
-            step_func=lambda: download_step(
-                titles=title_to_work,
-                output_dir=self.output_dir / "files",
+            if title_info.status.lower() in ["pending", "running"]:
+                title_info.status = "completed"
+
+            self.result.files_processed.append(title_info)
+
+            if ok and self.check_cancel_db_periodic():
+                logger.info("Job %s: Cancelled due to periodic check", self.job_id)
+                break
+
+            # Save progress after check for cancellation
+            if n == 1 or n % per_item == 0:
+                self._save_progress()
+
+        if processfiles_stage.status in ["pending", "running"]:
+            processfiles_stage.status = "completed"
+
+        return self.result
+
+    def inject_step_file(
+        self,
+        file_path_str: str,
+    ) -> Tuple[StepResult, None | str]:
+
+        if not file_path_str:
+            return StepResult(result=False, msg="No file path found"), None
+
+        file_path = Path(file_path_str)
+        output_file = self.output_dir / "translated" / file_path.name
+
+        inject_result: InjectResult = inject_step_one_file(
+            file_path,
+            self.translations,
+            output_file,
+            overwrite=self.overwrite_translations,
+        )
+
+        if inject_result.result is None:
+            step_result = StepResult(
+                result=None,
+                msg=inject_result.msg,
+            )
+            return step_result, None
+
+        if inject_result.result is False:
+            step_result = StepResult(
+                result=False,
+                msg=inject_result.msg,
+            )
+            return step_result, None
+
+        step_result = StepResult(
+            result=True,
+            msg=inject_result.msg,
+            details={
+                "new_languages": inject_result.new_languages,
+                "updated_translations": inject_result.updated_translations,
+            },
+        )
+        return step_result, output_file
+
+    def _process_one(self, title: str, title_info: FilesProcessedItem) -> bool:
+        self.result.summary.processed += 1
+
+        # ----------------------------------------------
+        # Stage 4: download SVG files
+
+        try:
+            download = download_svg_file(
+                title,
+                self.output_dir / "files",
                 session=self.session,
-                cancel_check=lambda: self._is_cancelled(self.result.stages.download),
-                overwrite_downloads=self.overwrite_downloads,
-                progress_callback=download_progress,
-            ),
-            run_after_func=download_run_after,
-        ):
-            return self.result
+            )
+        except Exception as e:
+            logger.exception("Error downloading SVG file")
+            title_info.steps.download = StepResult(result=False, msg="Error downloading", details={"error": str(e)})
+            title_info.status = "failed"
+            return False
+
+        if not download.get("ok"):
+            title_info.steps.download = StepResult(result=False, msg="Failed to download file", details=download)
+            title_info.status = "failed"
+            return False
+
+        title_info.steps.download = StepResult(result=True, msg="Downloaded successfully", details=download)
+
+        file_path = download.get("path")
+        if not file_path:
+            title_info.steps.download = StepResult(result=False, msg="Failed to get file path", details=download)
+            title_info.status = "failed"
+            return False
+
+        title_info.file_path = str(file_path)
+
         # ----------------------------------------------
         # Stage 5: Analyze And Fix Nested Files
 
-        def fix_nested_progress(index: int, total: int, msg: str, results: dict[str, Any]) -> None:
-            self.result.stages.nested.message = f"Analyzing {index}/{total}: {msg}"
-            if index % 10 == 0:
-                self._save_progress()
+        detect_before: DetectionResult = detect_nested_tags(file_path)
+        verify_fixed = 0
+        if detect_before.count == 0:
+            title_info.steps.nested = StepResult(result=None, msg="No nested tags found")
+        else:
+            # Try to fix nested tags
+            if not fix_nested_tags(file_path):
+                title_info.steps.nested = StepResult(
+                    result=False,
+                    msg="Failed to fix nested tags",
+                    details=detect_before.to_dict(),
+                )
+                title_info.status = "failed"
+                # We can't inject file that has nested tags
+                title_info.steps.inject.msg = "skipped"
+                title_info.steps.upload.msg = "skipped"
+                return False
 
-            for title, nested_result in results.items():
-                if title in self.result.files_processed:
-                    item = self.result.files_processed[title]
+            verify: VerificationResult = verify_fix(file_path, detect_before.count)
 
-                    if not item.steps.nested.msg:  # to avoid overwriting previous messages
-                        item.steps.nested = StepResult(
-                            result=nested_result.get("result", ""),
-                            msg=nested_result.get("msg", ""),
-                        )
-                        if nested_result.get("result", "") is False:
-                            item.status = "failed"
-                            item.error = nested_result.get("msg", "")
+            if verify.fixed == 0:
+                title_info.steps.nested = StepResult(
+                    result=False,
+                    msg="No nested tags were fixed",
+                    details=verify.to_dict(),
+                )
+                title_info.status = "failed"
+                # We can't inject file that has nested tags
+                title_info.steps.inject.msg = "skipped"
+                title_info.steps.upload.msg = "skipped"
+                return False
+            else:
+                verify_fixed = verify.fixed
 
-        def nested_run_after() -> None:
-            nested_stage_data = self.result.stages.nested.data
-            self.nested_data = nested_stage_data["data"]
-            self.nested_results = nested_stage_data.get("results", {})
-
-            # clean up
-            # self.result.stages.nested.data["data"] = {}
-            # self.result.stages.nested.data["results"] = {}
-
-            return
-
-        if not self._run_stage(
-            self.result.stages.nested,
-            step_func=lambda: fix_nested_step(
-                self.files_dict,
-                cancel_check=lambda: self._is_cancelled(self.result.stages.nested),
-                progress_callback=fix_nested_progress,
-            ),
-            run_after_func=nested_run_after,
-        ):
-            return self.result
         # ----------------------------------------------
+        # At this point, no nested tags remaining in the file
         # Stage 6: Inject translations
 
-        def inject_run_after() -> None:
-            inject_stage_data = self.result.stages.inject.data
+        inject_result, new_path = self.inject_step_file(title_info.file_path)
+        title_info.steps.inject = inject_result
 
-            inject_results = inject_stage_data["results"]
-            self.inject_data = inject_stage_data["data"]
+        if inject_result.result is True:
+            # inject success
+            new_languages = inject_result.details.get("new_languages", 0) if inject_result.details else 0
+            summary = (
+                f"Adding {new_languages} languages translations from {self.main_title}"
+                if new_languages > 0
+                else f"Adding translations from {self.main_title}"
+            )
+            # Stage 7: Upload
+            return self._upload_step(title_info, summary, new_path)
 
-            inject_files = self.inject_data.get("files", {})
-
-            self.files_to_upload = {title: data for title, data in inject_files.items() if data.get("file_path")}
-
-            # Update files_processed with inject results
-            for title, item in self.result.files_processed.items():
-                file_data = inject_results.get(title, {})
-
-                if not file_data:
-                    file_data = {"result": False, "msg": "Injection failed or skipped", "new_languages": 0}
-
-                item.steps.inject = StepResult(
-                    result=file_data.get("result"),
-                    msg=file_data.get("msg", ""),
-                )
-                if file_data.get("result") is False:
-                    item.status = "failed"
-                    item.error = file_data.get("msg", "")
-
-            # clean up
-            self.result.stages.inject.data["data"] = {}
-            self.result.stages.inject.data["results"] = {}
-
-        if not self._run_stage(
-            self.result.stages.inject,
-            step_func=lambda: inject_step(
-                self.files_dict,
-                self.translations,
-                self.output_dir,
-                overwrite=self.overwrite_translations,
-            ),
-            run_after_func=inject_run_after,
-        ):
-            return self.result
         # ----------------------------------------------
         # Stage 7: Upload
+        if verify_fixed > 0:
+            # Here we need to upload the orignal file because we fix nested tags.
+            summary = f"{verify_fixed} nested tags fixed"
+            return self._upload_step(title_info, summary, file_path)
 
-        def upload_progress(index: int) -> None:
-            if index % 10 == 0:
-                self._save_progress()
+        # No nested tags were fixed, and inject failed
+        return False
 
-        def upload_run_after() -> None:
-            upload_result_data = self.result.stages.upload.data
-            upload_result = {
-                "done": upload_result_data["summary"]["uploaded"],
-                "not_done": upload_result_data["summary"]["failed"],
-                "no_changes": upload_result_data["summary"]["no_changes"],
-                "errors": upload_result_data["errors"],
-            }
-            self.result.results_summary["upload_result"] = upload_result
-
-            upload_results = upload_result_data.get("results", {})
-
-            # Update files_processed with upload results
-            for title, item in self.result.files_processed.items():
-                if title in upload_results:
-                    item_upload_result = upload_results[title]
-                    item.steps.upload = StepResult(
-                        result=item_upload_result.get("result", ""),
-                        msg=item_upload_result.get("msg", ""),
-                    )
-
-                    if item_upload_result.get("result", "") is True:
-                        item.status = "completed"
-                    elif item_upload_result.get("result", "") is False:
-                        item.status = "failed"
-                        item.error = item_upload_result.get("msg", "")
-                    elif item_upload_result.get("result", "") is None:
-                        # Skipped or no changes
-                        item.status = "completed"
-
-                if item.status == "pending":
-                    item.status = "completed"
-
+    def _upload_step(
+        self,
+        title_info: FilesProcessedItem,
+        summary: str,
+        new_path: Path,
+    ) -> bool:
+        # Check if upload is enabled
         if not bool(self.args.get("upload")):
-            upload_result = {"done": 0, "not_done": len(self.files_to_upload), "skipped": True}
-            self.result.results_summary["upload_result"] = upload_result
-            self.log_upload_error("Upload disabled", None, "Skipped")
-            # ----------------------------------------------
-            # Stage 8: save stats and mark done
-            self.save_stats_summary()
+            title_info.steps.upload = StepResult(
+                result=None,
+                msg="skipped",
+                details={"error": "Upload disabled"},
+            )
+            title_info.status = "skipped"
+            return False
 
-            return self.result
+        if self.upload_limit > 0 and self.upload_done >= self.upload_limit:
+            title_info.steps.upload = StepResult(
+                result=None,
+                msg="skipped",
+                details={"error": "Upload limit reached"},
+            )
+            title_info.status = "skipped"
+            return False
 
         # Start uploading
-        if not self._run_stage(
-            self.result.stages.upload,
-            step_func=lambda: upload_step(
-                self.files_to_upload,
-                self.main_title,
-                self.site,
-                cancel_check=lambda: self._is_cancelled(self.result.stages.upload),
-                progress_callback=upload_progress,
-                upload_limit=self.upload_limit,
-            ),
-            run_after_func=upload_run_after,
-        ):
-            return self.result
+        upload = upload_fixed_svg(
+            title_info.title,
+            new_path,
+            0,
+            self.site,
+            summary=summary,
+        )
+        upload_success = upload.get("ok")
+        upload_error = upload.get("error") or ""
+        upload_msg = upload.get("msg") or ""
 
-        # ----------------------------------------------
-        # Stage 8: save stats and mark done
-        self.save_stats_summary()
+        if upload_success is True:
+            title_info.steps.upload = StepResult(
+                result=True,
+                msg="File Successfully uploaded.",
+                # details=upload.get("result", ""),
+            )
 
-        return self.result
+            self.upload_done += 1
+            title_info.status = "completed"
+            # return True, all steps passed and upload is success
+            return True
+
+        error_and_details = {
+            "error": upload_error,
+            "error_details": upload.get("error_details", ""),
+        }
+
+        if upload_success is None and upload_error == "skipped":
+            title_info.steps.upload = StepResult(
+                result=None,
+                msg=upload_msg,
+                details=error_and_details,
+            )
+            return False
+
+        title_info.error = upload_error
+        title_info.steps.upload = StepResult(
+            result=False,
+            msg="Upload failed.",
+            details=error_and_details,
+        )
+        return False
 
 
 # --- main pipeline --------------------------------------------
