+++ src/templates/_navbar.html
@@ -26,8 +26,8 @@
                             'public_jobs.jobs_list',
                             'bi-list-task',
                             'All Copy Tasks',
-                            url_for('public_jobs.jobs_list', job_type='copy_svg_langs_per_file' ),
-                            "/jobs/copy_svg_langs_per_file"
+                            url_for('public_jobs.jobs_list', job_type='copy_svg_langs' ),
+                            "/jobs/copy_svg_langs"
                             )
                         }}
                     </li>
