+++ src/templates/index.html
@@ -5,7 +5,7 @@
 {% block content_fluid %}
 <div class="container">
     <h1 class="mb-3">{{ tool_title }}</h1>
-    <form method="post" action="{{ url_for('public_jobs.start_job', job_type='copy_svg_langs_per_file') }}"
+    <form method="post" action="{{ url_for('public_jobs.start_job', job_type='copy_svg_langs') }}"
         class="row gy-2 gx-3 align-items-center mb-4">
         <input type="hidden" name="csrf_token" value="{{ csrf_token() }}" />
         <div class="col-sm-12 col-md-4">
