--- src/templates/jobs_templates/public/copy_svg_langs_per_file/details.html
@@ -1,166 +1,0 @@
-{% extends "jobs_templates/base_details_public.html" %}
-{% from "_macros.html" import card_header, stats_card, status_icon, commons_file_link %}
-
-{% block job_info_extra %}
-{% if result_data and result_data.title %}
-<tr>
-    <th>Main Title:</th>
-    <td>
-        <a href="https://commons.wikimedia.org/wiki/{{ result_data.title }}" target="_blank" rel="noopener noreferrer">
-            {{ result_data.title }}
-        </a>
-    </td>
-</tr>
-{% endif %}
-{% endblock %}
-
-{% block job_summary %}
-
-<!-- Pipeline Stages -->
-{% if result_data.stages %}
-<div class="table-responsive">
-    <table class="table table-striped table-sm mb-0">
-        <thead>
-            <tr>
-                <th>#</th>
-                <th>Stage</th>
-                <th>Status</th>
-                <th>Message</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% set stage_order = ['text', 'titles', 'translations', 'processfiles'] %}
-            {% for stage_key in stage_order %}
-            {% set stage = result_data.stages.get(stage_key) %}
-            {% if stage %}
-            <tr>
-                <td> {{ loop.index }} </td>
-                <td>{{ stage_key|title }}</td>
-                <td>
-                    {{ status_icon(stage.status) }}
-                </td>
-                <td><small>{{ stage.message }}</small></td>
-            </tr>
-            {% endif %}
-            {% endfor %}
-        </tbody>
-    </table>
-</div>
-{% endif %}
-{% endblock %}
-
-{% block job_details %}
-
-
-<!-- Files Processed -->
-{% if result_data.files_processed %}
-<div class="card mb-3">
-    <div class="card-header">
-        {{ card_header("Files Processed (" ~ result_data.files_processed|length ~ ")") }}
-    </div>
-    <div class="card-body">
-        <div class="table-responsive">
-            <table class="table table-striped table-sm">
-                <thead>
-                    <tr>
-                        <th>#</th>
-                        <th>File</th>
-                        <th>Status</th>
-                        <th>Download</th>
-                        <th>Nested</th>
-                        <th>Inject</th>
-                        <th>Upload</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {% set step_keys = [ 'download', 'nested', 'inject', 'upload' ] %}
-                    {% for item in result_data.files_processed %}
-                    <tr>
-                        <td> {{ loop.index }} </td>
-                        <td>
-                            {% set short_title = item.title[:50] ~ '...' if item.title|length > 50 else item.title %}
-                            {{ commons_file_link(item.title, label=short_title) }}
-                        </td>
-                        <td>
-                            {{ status_icon(item.status) }}
-                        </td>
-                        {% for step_key in step_keys %}
-                        <td>
-                            {% set step = item.steps.get(step_key) if item.steps else None %}
-                            {% if step %}
-                            {% if step.result == true %}
-                            <span class="badge bg-success" title="{{ step.msg }}">
-                                <i class="bi bi-check-lg"></i>
-                            </span>
-                            {% elif step.result == false %}
-                            <span class="badge bg-danger" title="{{ step.msg }}">
-                                <i class="bi bi-x-lg"></i>
-                            </span>
-                            {% elif step.result is none %}
-                            <span class="badge bg-secondary" title="{{ step.msg }}">
-                                <i class="bi bi-dash"></i>
-                            </span>
-                            {% else %}
-                            <span class="text-muted">-</span>
-                            {% endif %}
-                            {% else %}
-                            <span class="text-muted">-</span>
-                            {% endif %}
-                        </td>
-                        {% endfor %}
-                    </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
-        </div>
-    </div>
-</div>
-{% endif %}
-
-<!-- Translations -->
-{% if result_data.stages and result_data.stages.translations and result_data.stages.translations.data %}
-{% set trans_data = result_data.stages.translations.data.translations %}
-{% if trans_data and trans_data.new %}
-{% set new_trans = trans_data.new %}
-{% set new_count = new_trans | length %}
-<div class="card mb-3">
-    <div class="card-header">
-        {{ card_header("New Translations (" ~ new_count ~ ")") }}
-    </div>
-    <div class="card-body">
-        <div class="table-responsive">
-            <table class="table table-striped table-sm">
-                <thead>
-                    <tr>
-                        <th>#</th>
-                        <th>Original Text</th>
-                        {% set first_key = new_trans | dictsort | first %}
-                        {% if first_key %}
-                        {% set first_val = new_trans[first_key[0]] %}
-                        {% for lang in first_val %}
-                        <th>{{ lang }}</th>
-                        {% endfor %}
-                        {% endif %}
-                    </tr>
-                </thead>
-                <tbody>
-                    {% for key, translations in new_trans | dictsort %}
-                    {% if translations %}
-                    <tr>
-                        <td> {{ loop.index }} </td>
-                        <td><small class="text-muted">{{ key[:80] }}{% if key|length > 80 %}...{% endif %}</small></td>
-                        {% for lang, text in translations.items() %}
-                        <td><small>{{ text[:80] }}{% if text|length > 80 %}...{% endif %}</small></td>
-                        {% endfor %}
-                    </tr>
-                    {% endif %}
-                    {% endfor %}
-                </tbody>
-            </table>
-        </div>
-    </div>
-</div>
-{% endif %}
-{% endif %}
-
-{% endblock %}
